Stores doubles

+++ jpa/src/main/java/io/freedriver/autonomy/jpa/converter/measurement/MeasurementConverter.java
@@ -6,18 +6,19 @@ import kaze.math.measurement.types.Measurement;
 import javax.persistence.AttributeConverter;
 import java.math.BigDecimal;
 
-public abstract class MeasurementConverter<M extends Measurement<M>> implements AttributeConverter<M, BigDecimal> {
+public abstract class MeasurementConverter<M extends Measurement<M>> implements AttributeConverter<M, Double> {
     public abstract M construct(BigDecimal value, UnitPrefix unitPrefix);
 
     @Override
-    public BigDecimal convertToDatabaseColumn(M m) {
+    public Double convertToDatabaseColumn(M m) {
         return m.scaleTo(UnitPrefix.ONE).getValue()
-                .getValue();
+                .getValue()
+                .doubleValue();
     }
 
     @Override
-    public M convertToEntityAttribute(BigDecimal value) {
-        return construct(value, UnitPrefix.ONE)
+    public M convertToEntityAttribute(Double value) {
+        return construct(BigDecimal.valueOf(value), UnitPrefix.ONE)
                 .normalize();
     }
 }
